Skill Calc plugin redesign
Following the obsidian redesign update, this time I've ventured into
the skilling calculators. Major changes were made, here are the most
noticeable:

- Added a title to the top of the plugin panel
- Resized the skilling selectors to a 3x6 grid and used the MaterialTabs
for the selection behavior.
- Resized, recolored and reshaped the input panels and respective labels
to better fit the new theme and aesthetic.
- Added a new action "on enter", the textfield will now request focus
on the next logical textfield, (example, when enter is pressed on the
current level field, it will request focus on the target level)
- Reworded the input labels from "XP" to "Experience"
- Resized, recolored and reshaped the boost/modifier selectors
- Recolored the actions/selections indicator to a dark gray
- Recolored all item panels to a dark gray, and instead added a left
side thick border to indicate availability (green/red)
- Added required level indicator on the sub-title of each item
- Added hover effects for the individual items
- Added hover effects for the material tabs