Add solutions for 3 step cryptic master clues
This extends the functionality of the clue scroll plugin to include
solutions for three step cryptic master clues. The solutions for these
clues are already present in the existing CrypticClue class, but before
this there was not a way for the clues to be recognized. The
ThreeStepCrypticClue class matches the clue scroll text for all three
cryptic clues and keeps the matched clue objects as a list, along with
their completion status. If a clue has a strikethrough tag when reading
the clue, then the plugin will recognize it as complete. The plugin will
also detect when a new clue part has been obtained and update the
completed status of that clue.