itemstats: Fix unarmed attack speed
Since the writing of this plugin, the wiki has changed the format for
weapon attack speeds[1] and now displays weapon attack speeds in length
of game ticks. Hence, since unarmed attacks are at the rate of once
every 4 game ticks, this commit updates our definition of the unarmed
attack accordingly.

[1] https://oldschool.runescape.wiki/w/Unarmed?diff=8467472&oldid=7810087