fishing: Check player animation to update fishing status
This commit adds an animation check to the other fishing status checks
already present in the fishing overlay (interacting and fishing spot NPC
name check) to ensure that the status correctly changes to "NOT fishing"
when the player's fishing animation is interrupted by certain actions
such as being presented with a level-up dialog. This commit causes one
other behavioral change as a side effect, which is that the player's
fishing status will be "NOT fishing" while running toward a distant
fishing spot until they reach it and start their fishing animation.

Fixes runelite/runelite#10900