loottracker: Track herbiboar loot with an open herb sack
When looting the herbiboar with an open herb sack, instead of receiving
the herbs to your inventory, the herbs are deposited to the sack
directly and the player receives chat messages indicating the herbs they
received. This commit adds support for reading those chat messages when
looting the herbiboar.

This change incidentally fixes an issue which arose when the open herb
sack was introduced to the game where any inventory change after looting
the herbiboar would be tracked as herbiboar loot, such as drinking a
stamina potion.

Closes #10655
Fixes #10718