key remapper: consume key typed events for remapped keys
Since all of the keys we remap to (f/arrow/escape keys) do not have
associated characters, consume key typed events for the characters
associated to the original key press event.

This fixes entering bank pins with fkey remapping on both entering your
pin and changing your inventory tab, due to the client processing the
tab change on key press (which is remapped) and the bank plugin
processing the pin input on key typed.