cluescrolls: Fix beginner/master clue step reset
Commit 5a3d4effbff9165e2d4354715e9059affde0515a improved the clue scroll
plugin by resetting the active clue step when a beginner or master clue
item was displayed in the chat dialog when receiving a new step. This
commit caused a new bug, however, in that it became possible to click on
a new clue step while the new step dialog was open such that the new
step dialog persisted through a GameTick due to client-server delay,
causing the saved clue item ID to be lost. This commit fixes that issue
by resetting only the active clue object, but not the recorded clue item
ID, as beginner and master clues only use a single item ID.