key remapping: set press enter to chat text regardless of chatbox focus
I can't figure out why this was originally added, but this shouldn't
care about the chatbox focus state since the script event is in the
script which builds the chatbox input widget. This is currently causing
an issue where the input rebuild script runs immediately before the
chatbox input keylistener is reapplied when closing input dialogs.