banktags: Fix bank scroll height with separators hidden (#13723)
Prior to this change, an extra row of empty space was added to tag tabs
with items reaching the end of a row (ie. multiples of 8) which led to
scroll behavior not matching vanilla for tabs with enough items to
require scrolling. This change fixes this behavior to add extra scroll
height only when a new row is started.

In addition to the 36 pixels of height allotted to each bank item row,
an additional 4 pixels of height exists at the bottom of the bank item
container. This change ensures that a bank tag scrolled to the bottom
will have its items displaying at the exact same y values as that of a
normal bank tab.