Fix race with client startup and low detail/chat icon loading
Since 835b63d341d0180b636e4f2251b7c34859065c07 moved client startup to
be much earlier, the plugins can now start after the client is already
at the login screen. This causes the low detail plugin and chat icon
manager to not initialize correctly as they are waiting on the event
for the login screen.