widget overlay: use setForcedPosition
The inventory can be moved pre-interface tick by the tli subchange
listener, as well as post-our client tick event from script events
running.

This makes it difficult to correctly and generically move the interface,
since it would have to be moved pre-interface tick and also pre-frame.

Currently the code moves the interface pre-frame, which looks okay, but
since it is ticking the interface when it is a different position,
clicks on it do not work correctly.

We do not have events for pre-interface tick, so use forced position to
reliably set the position