item charges: fix tracking explorer ring charges
ce94421a2274ab67191e8f348abae1c99ca5ae0b no longer causes the varbit to
be checked on login due to other varbits changing. If the value is 0,
for full charges, no event will be fired because the default varbit
value is 0.

This initializes the charges to 0 on login to handle this case