idle notifier: Re-check NPC interaction on transform
NPCs which begin combat with a menu option that is not "Attack" are
missed during combat detection. This commit uses NpcChanged to detect
when an NPC that we are interacting with transforms and gains an Attack
option while still being interacted with. This notably fixes the
interaction check at Kraken as combat is initiated via a "Disturb"
option on a Whirlpool which transforms into the Kraken.

Co-authored-by: Jordan Atwood <nightfirecat@nightfirec.at>