gpu: fix amd miscompile of frag.glsl
amd drivers 22.7.1 and later miscompile a switch with two leading cases when one falls through.

The switch was originally put in place due to intel drivers breaking with a chain of ifs, which was fixed in a4b497c24d653c6668a4c69342be5c8b31e8f4cd.

Hopefully this works better since there is an else {} case and also the color blind code is not behind a condition anymore.