cache: modify Storage api to accept int index/archive
This more closely matches the client load/store behavior which just
indexes into files.

Previously this was also setting the archive crc in store, which seems
wrong. It no longer does this, and so callers need to ensure they have
the correct crc set.

compress() now updates the container crc to make this easier.