gpu: remove sin/cos table from UBO
This was added in 3bbcfccc09032108c703029f0c5805fddef51a34 due to
inconsistencies in the computed sin/cos values causing graphical issues
on some cards. However the sin/cos table is no longer used for the
projection, and instead only for computing face distance for face
sorting, so I think it is safe to remove.