config manager: Fix possible profile display name NPE
Under very specific and unlikely scenarios, it is possible for a user to
have an RSProfile with no display name saved to it. In this case, a user
launching the client via the Jagex launcher may hit the profile and
launcher display name comparison with the profile display name being
null. Because the launcher display name is already checked to be
non-null, flipping the comparison is guaranteed to be a safe operation.