gpu: fix priority face average calculation
While the client defaults the averages to 0, it also offsets the
distances by the model radius, meaning 0 is always the closest distance.

This is a regression from c501b3c8623a49a71c6fadf85e84c40ebf5efeee which
removed the radius offset from the distances.

Initialize averages to -6000 because the model max diameter is 6000 and
so it should always be less than the computed distance.